refactor(memory): split store.ts into store/staleness/enrichment
store.ts was 747 lines doing four jobs: raw SQL CRUD, entity
resolution, per-proposition provenance staleness checking, and
read-side query helpers (neighbors, counts, status). The review
flagged it as approaching the size where every change requires
re-orientation, and surfaced a separate concern: the provenance
caches (mtimeCache, hashCache) were instance-level and grew
unbounded for the lifetime of the store.

Three-way split with a deliberate cache-scope fix:

1. **src/memory/staleness.ts** (new, 120 lines) — owns StalenessCache,
   isFileChanged, getStalePropositionIds. The cache is now a per-call
   scratchpad: every public read on MemoryStore creates a fresh cache
   with createStalenessCache(), threads it through staleness checks
   to dedupe stat()/read() within that one operation, and lets it
   garbage-collect when the call returns. No more instance-level
   caches, no unbounded growth, no explicit eviction needed — the
   cache scope matches the working set of a single query.

2. **src/memory/enrichment.ts** (new, 162 lines) — owns the read-side
   query helpers (collectionFilter, getNeighbors, countValidForEntity,
   computeStatus) as pure functions. Each takes a Db handle, any
   pre-computed stalePropIds the caller has, and optional collection
   filter. No hidden state.

3. **src/memory/store.ts** (567 lines, down from 747) — keeps the
   public API, constructor/config, entity resolution (resolveEntity,
   findEntity), source path resolution, registerSource, emit, and
   enrichProposition (the private helper every read uses — kept on
   the class because moving it to enrichment.ts would mean threading
   sourceRoot + cache through every call site in a file where it's
   the local concern). Public read methods (browse, inspect, bySource,
   search, status, related) now create a cache, compute staleness via
   the new module, and delegate query logic to enrichment helpers.

Behavior is identical: same SQL, same results, same staleness
semantics. The 576/576 tests pass unchanged. Biome clean.

The separation also makes future changes easier: adding a new query
helper is a function in enrichment.ts, not another method wedged into
MemoryStore. Changing the mtime vs hash fallback strategy is localized
to staleness.ts. The SQL CRUD and workflow semantics stay in store.ts
where they belong.
